feature: adds code reorganization for vercel deployment
